A02_Cryptographic_Failures/environment/requirements.txt
Flask==2.3.3
gunicorn==21.2.0
PyJWT==2.8.0

# Flask 依赖（必须显式写）
Werkzeug==3.1.7
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.1.8
blinker==1.9.0
MarkupSafe==3.0.3

# gunicorn 依赖
packaging==26.0

importlib-metadata==6.8.0
zipp==3.20.2